requirements.txt
astropy
cython
numpy
pulsar_spectra
plotly
pandas
psrqpy
poliastro
pint-pulsar
scipy
